fix silly warnings

+++ tinyshell/libtinyshell/utils.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <assert.h>
 #include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -38,7 +37,7 @@ inline static int vecpush(void * /* T** */ dst, int *len, int *cap,
   void **dest = (void **)dst;
   const float scale_factor = 1.5;
   if (*len + srclen >= *cap) {
-    int new_cap = max_int(*cap * scale_factor, *len + srclen);
+    int new_cap = max_int((int)(*cap * scale_factor), *len + srclen);
     void *new_dest = realloc(*dest, new_cap * elemsize);
     if (!new_dest) {
       return 0;
@@ -71,6 +70,6 @@ inline static char *reentrant_strtok(char *s, const char *sep, char **p) {
 }
 
 inline static int string_ends_with(const char *str, const char *suffix) {
-  int str_len = strlen(str), suf_len = strlen(suffix);
+  size_t str_len = strlen(str), suf_len = strlen(suffix);
   return str_len >= suf_len && strcmp(str + str_len - suf_len, suffix) == 0;
 }
